Convert cost calculation to long instead of int after rounding in status bar.
This was an issue for some dropships that were showing a negative cost
in the status bar, but I went ahead and made the change for all units.

+++ src/megameklab/com/ui/Aero/StatusBar.java
@@ -108,7 +108,7 @@ public class StatusBar extends ITab {
         double tonnage = getAero().getWeight();
         double currentTonnage;
         int bv = getAero().calculateBattleValue();
-        int currentCost = (int)Math.round(getAero().getCost(false));
+        long currentCost = (long) Math.round(getAero().getCost(false));
 
         testAero = new TestAero(getAero(), entityVerifier.aeroOption, null);
